home: only jump to end-of-page when it actually nears
The main reason of this workaround is to keep error message fully displayed
to the users since the height of previous view (loading bar) are pretty
small compared to the error view, and we have to make a workaround for
the view to fully visible instead of user having to scroll manually.

But there will be a potential case where user did not yet reach the view
and user will be forced to scroll to the floor by our cheap workaround.