Introduce PostHeadersFactory
Using helper method to build a simple headers was fine, until it feels
like we're fighting with the tool itself (which in case, provider
reference between normal provider usage and directly from the widget).

So instead of using helper method, we can just create factory class that
already holds the necessary dependencies to build headers when needed
and then expose the build method for users.