Construct needed headers for server who uses json
Szurubooru need proper Accept and Content-Type header in order to
request to their API. So let's construct it when query has json=1.

Additionally, we can also enforce it from URL fragment as well in case
that we don't want to expose json=1 to the server by adding "#opt?json=1"
to the end of the query.

Issue: #121